# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Temporary Access Manager API
  description: API para cadastro de usuários e gerenciamento de acessos temporários.
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Acessos
  description: "Concessão, consulta e revogação de acessos temporários"
- name: Autenticação
  description: Cadastro e autenticação de usuários
- name: Usuários
  description: Administração de usuários
paths:
  /api/usuario/rejeitar/{id}:
    post:
      tags:
      - Usuários
      summary: Rejeita um usuário pendente
      description: "Endpoint protegido. Perfil necessário: ADMIN."
      operationId: rejeitar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: Usuário rejeitado
          content:
            text/plain:
              example: Rejeitado
        "400":
          $ref: "#/components/responses/EntradaInvalida"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
      security:
      - bearerAuth: []
  /api/usuario/aprovar/{id}:
    post:
      tags:
      - Usuários
      summary: Aprova um usuário pendente
      description: "Endpoint protegido. Perfil necessário: ADMIN."
      operationId: aprovar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: Usuário aprovado
          content:
            text/plain:
              example: Aprovado
        "400":
          $ref: "#/components/responses/EntradaInvalida"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
      security:
      - bearerAuth: []
  /api/auth/registrar:
    post:
      tags:
      - Autenticação
      summary: Cadastra um usuário
      description: Endpoint público. Cria um usuário com status pendente de aprovação.
      operationId: registrar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroRequestDto"
        required: true
      responses:
        "200":
          description: Usuário cadastrado
          content:
            text/plain:
              example: Registrado. Esperando para ser aprovado.
        "400":
          $ref: "#/components/responses/EntradaInvalida"
        "409":
          $ref: "#/components/responses/Conflito"
  /api/auth/login:
    post:
      tags:
      - Autenticação
      summary: Autentica um usuário
      description: Endpoint público. Retorna um token JWT para usuários aprovados.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequestDto"
        required: true
      responses:
        "200":
          description: Autenticação realizada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
  /api/acesso/revogar/{idAcesso}:
    post:
      tags:
      - Acessos
      summary: Revoga um acesso
      description: "Endpoint protegido. Perfil necessário: ADMIN."
      operationId: revogar
      parameters:
      - name: idAcesso
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: Acesso revogado
          content:
            text/plain:
              example: Revogado
        "400":
          $ref: "#/components/responses/EntradaInvalida"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
      security:
      - bearerAuth: []
  /api/acesso/permitir:
    post:
      tags:
      - Acessos
      summary: Concede um acesso temporário
      description: "Endpoint protegido. Perfil necessário: ADMIN."
      operationId: permitir
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissaoAcessoRequestDto"
        required: true
      responses:
        "200":
          description: Acesso concedido
          content:
            application/json:
              example: true
        "400":
          $ref: "#/components/responses/EntradaInvalida"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
      security:
      - bearerAuth: []
  /api/usuario:
    get:
      tags:
      - Usuários
      summary: Lista os usuários aprovados
      description: "Endpoint protegido. Perfil necessário: ADMIN."
      operationId: listAllUsuarios
      responses:
        "200":
          description: Usuários encontrados
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsuarioResponseDto"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
      security:
      - bearerAuth: []
  /api/usuario/pendentes:
    get:
      tags:
      - Usuários
      summary: Lista os usuários pendentes
      description: "Endpoint protegido. Perfil necessário: ADMIN."
      operationId: listarPendentes
      responses:
        "200":
          description: Usuários pendentes encontrados
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsuarioResponseDto"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
      security:
      - bearerAuth: []
  /api/acesso:
    get:
      tags:
      - Acessos
      summary: Lista todos os acessos
      description: "Endpoint protegido. Perfil necessário: ADMIN."
      operationId: getAllAcessos
      responses:
        "200":
          description: Acessos encontrados
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcessoResponseDto"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
      security:
      - bearerAuth: []
  /api/acesso/usuario:
    get:
      tags:
      - Acessos
      summary: Lista os acessos do usuário autenticado
      description: "Endpoint protegido. Perfil necessário: USER."
      operationId: listAcessosUser
      responses:
        "200":
          description: Acessos do usuário encontrados
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcessosUserResponseDto"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
      security:
      - bearerAuth: []
components:
  schemas:
    RegistroRequestDto:
      type: object
      properties:
        nome:
          type: string
          example: Maria Silva
          minLength: 1
        email:
          type: string
          format: email
          example: maria@example.com
          minLength: 1
        senha:
          type: string
          example: senha-segura
          minLength: 1
      required:
      - email
      - nome
      - senha
    AuthRequestDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: usuario@example.com
          minLength: 1
        senha:
          type: string
          example: senha-segura
          minLength: 1
      required:
      - email
      - senha
    AuthResponseDto:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        role:
          type: string
          enum:
          - ADMIN
          - USER
          example: USER
    PermissaoAcessoRequestDto:
      type: object
      properties:
        usuarioId:
          type: integer
          format: int64
          example: 1
        nomeRecurso:
          type: string
          example: Painel financeiro
          minLength: 1
        duracaoSegundos:
          type: integer
          format: int64
          description: "Duração em segundos, limitada a 30 dias"
          example: 3600
          maximum: 2592000
      required:
      - duracaoSegundos
      - nomeRecurso
      - usuarioId
    UsuarioResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: Maria Silva
        email:
          type: string
          example: maria@example.com
    AcessoResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        nomeRecurso:
          type: string
          example: Painel financeiro
        horaPermissao:
          type: string
          format: date-time
          example: 2026-08-16T12:00:00Z
        horaExpiracao:
          type: string
          format: date-time
          example: 2026-08-16T13:00:00Z
        revogado:
          type: boolean
          example: false
        usuarioId:
          type: integer
          format: int64
          example: 1
        nomeUsuario:
          type: string
          example: Maria Silva
    AcessosUserResponseDto:
      type: object
      properties:
        nomeRecurso:
          type: string
          example: Painel financeiro
        horaPermissao:
          type: string
          format: date-time
          example: 2026-08-16T12:00:00Z
        horaExpiracao:
          type: string
          format: date-time
          example: 2026-08-16T13:00:00Z
  responses:
    EntradaInvalida:
      description: Entrada inválida
      content:
        application/problem+json:
          examples:
            exemplo:
              value:
                title: Entrada inválida
                instance: /api/exemplo
                type: https://example.com/errors/exemplo
                status: 400
                detail: Entrada inválida.
    NaoAutenticado:
      description: "Token ausente, inválido ou credenciais incorretas"
      content:
        application/problem+json:
          examples:
            exemplo:
              value:
                title: "Token ausente, inválido ou credenciais incorretas"
                instance: /api/exemplo
                type: https://example.com/errors/exemplo
                status: 401
                detail: "Token ausente, inválido ou credenciais incorretas."
    SemPermissao:
      description: Perfil sem permissão para esta operação
      content:
        application/problem+json:
          examples:
            exemplo:
              value:
                title: Perfil sem permissão para esta operação
                instance: /api/exemplo
                type: https://example.com/errors/exemplo
                status: 403
                detail: Perfil sem permissão para esta operação.
    NaoEncontrado:
      description: Usuário ou acesso não encontrado
      content:
        application/problem+json:
          examples:
            exemplo:
              value:
                title: Usuário ou acesso não encontrado
                instance: /api/exemplo
                type: https://example.com/errors/exemplo
                status: 404
                detail: Usuário ou acesso não encontrado.
    Conflito:
      description: Conflito com o estado atual do recurso
      content:
        application/problem+json:
          examples:
            exemplo:
              value:
                title: Conflito com o estado atual do recurso
                instance: /api/exemplo
                type: https://example.com/errors/exemplo
                status: 409
                detail: Conflito com o estado atual do recurso.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT